chore: move TypeInstrumention into standalone class
